feat(admin): add LabTableView component
