+++ apps/admin/src/components/lab/lab-table.tsx
@@ -1,17 +1,8 @@
-import { UploadOutlined } from '@ant-design/icons';
-import type { TableProps } from 'antd';
-import {
-  Button,
-  Form,
-  Input,
-  Popconfirm,
-  Table,
-  Typography,
-  Upload,
-} from 'antd';
+import { Form } from 'antd';
 import { useState } from 'react';
+import LabTableView from './lab-table-view';
 
-interface LabData {
+export interface LabData {
   key: string;
   name: string;
   loc: string;
@@ -33,7 +24,7 @@ const labData: LabData[] = [
     img: {
       id: 1,
       physicalPath:
-        'https://aics-client.vercel.app/_next/image?url=https%3A%2F%2Fwww.kyonggi.ac.kr%2Fsite%2Fu_ai%2Fimages%2Fcontents%2Fkey9123_img01.png&w=3840&q=75',
+        'https://www.kyonggi.ac.kr/site/u_ai/images/contents/key9123_img01.png',
     },
   },
   {
@@ -48,39 +39,6 @@ const labData: LabData[] = [
     },
   },
 ];
-
-interface EditableCellProps extends React.HTMLAttributes<HTMLElement> {
-  editing: boolean;
-  dataIndex: string;
-  title: string;
-  record: LabData;
-  index: number;
-}
-
-function EditableCell({
-  editing,
-  dataIndex,
-  title,
-  record,
-  children,
-  ...restProps
-}: EditableCellProps) {
-  return (
-    <td {...restProps}>
-      {editing ? (
-        <Form.Item
-          name={dataIndex}
-          style={{ margin: 0 }}
-          rules={[{ required: true, message: `${title}을(를) 입력해주세요!` }]}
-        >
-          <Input />
-        </Form.Item>
-      ) : (
-        children
-      )}
-    </td>
-  );
-}
 
 function LabTable() {
   const [form] = Form.useForm();
@@ -111,114 +69,16 @@ function LabTable() {
     // TODO: 연구실 삭제 DELETE API 연동
   };
 
-  const handleImageUpload = (file: File) => {
-    const reader = new FileReader();
-    reader.readAsDataURL(file);
-  };
-
-  const columns = [
-    { title: '연구실명', dataIndex: 'name', width: '15%', editable: true },
-    { title: '위치', dataIndex: 'loc', width: '15%', editable: true },
-    {
-      title: '웹사이트',
-      dataIndex: 'site',
-      width: '15%',
-      editable: true,
-      render: (text: string) => (
-        <a href={text} target="_blank" rel="noopener noreferrer">
-          {text}
-        </a>
-      ),
-    },
-    { title: '담당교수', dataIndex: 'advisor', width: '10%', editable: true },
-    {
-      title: '관리',
-      dataIndex: 'operation',
-      width: '10%',
-      render: (_: unknown, record: LabData) => {
-        const editable = isEditing(record);
-        return editable ? (
-          <span>
-            <Typography.Link onClick={() => handleSave()} className="mr-4">
-              저장
-            </Typography.Link>
-            <Popconfirm title="정말 취소하시겠습니까?" onConfirm={cancel}>
-              <button type="button" className="text-red-500">
-                취소
-              </button>
-            </Popconfirm>
-          </span>
-        ) : (
-          <span>
-            <Typography.Link
-              disabled={editingKey !== ''}
-              onClick={() => handleEdit(record)}
-            >
-              수정
-            </Typography.Link>
-            <Popconfirm
-              title="정말 삭제하시겠습니까?"
-              onConfirm={() => handleDelete()}
-            >
-              <button type="button" className="ml-4 text-red-500">
-                삭제
-              </button>
-            </Popconfirm>
-          </span>
-        );
-      },
-    },
-    {
-      title: '연구실 이미지',
-      dataIndex: 'img',
-      width: '20%',
-      render: (_: unknown, record: LabData) => (
-        <div>
-          {record.img?.physicalPath && (
-            <img
-              src={record.img.physicalPath}
-              alt={`${record.name} 이미지`}
-              className="w-24 h-24"
-            />
-          )}
-          <Upload
-            showUploadList={false}
-            beforeUpload={(file) => handleImageUpload(file)}
-          >
-            <Button icon={<UploadOutlined />} className="mt-2">
-              업로드
-            </Button>
-          </Upload>
-        </div>
-      ),
-    },
-  ];
-
-  const mergedColumns: TableProps<LabData>['columns'] = columns.map((col) => {
-    if (!col.editable) return col;
-
-    return {
-      ...col,
-      onCell: (record: LabData) => ({
-        record,
-        dataIndex: col.dataIndex,
-        title: col.title,
-        editing: isEditing(record),
-      }),
-    };
-  });
-
   return (
-    <Form form={form} component={false}>
-      <Table<LabData>
-        components={{ body: { cell: EditableCell } }}
-        bordered
-        dataSource={labData}
-        columns={mergedColumns}
-        rowClassName="editable-row"
-        pagination={{ onChange: cancel }}
-      />
-    </Form>
+    <LabTableView
+      form={form}
+      data={labData}
+      isEditing={isEditing}
+      handleEdit={handleEdit}
+      handleSave={handleSave}
+      handleDelete={handleDelete}
+      cancel={cancel}
+    />
   );
 }
